修复bug，提交部分测试用例

+++ chapter_05/cb/strategy.go
@@ -4,7 +4,7 @@ type Strategy interface {
 	GetRecoveryTime() int     //N秒后尝试恢复
 	GetMinFailCount() int     //最小失败次数
 	GetFailRate() float64     //单位%
-	GetTryCount() uint32      //尝试恢复的请求数量
+	GetTryCount() int64       //尝试恢复的请求数量
 	GetRecoveryRate() float64 //尝试请求成功率超过N，会由半开恢复为关闭，单位%
 }
 
@@ -23,7 +23,7 @@ func (receiver DefaultStrategy) GetFailRate() float64 {
 	return 30
 }
 
-func (receiver DefaultStrategy) GetTryCount() uint32 {
+func (receiver DefaultStrategy) GetTryCount() int64 {
 	return 5
 }
